feat: image slider components create
image slider components 메인, 게시물 모달에 사용
