+++ components/modal/CommonModal.module.scss
@@ -126,7 +126,7 @@
 @media (min-width: 940px) {
   .innerContainer {
     display: grid;
-    width: 860px;
+    width: 890px;
     grid-template-columns: 2fr 1fr;
   }
