+++ components/ui/ImageSlider/index.ts
@@ -1,0 +1,1 @@
+export { default as ImageSlider } from './ImageSlider'
